fix: 为大图模式下的右键菜单删除和下方工具栏删除添加确认界面
直接在两个对应位置呼出图片删除确认界面，确认即按照之前的删除步骤进行，取消则不进行删除

Log: 为大图模式下的右键菜单删除和下方工具栏删除添加确认界面

+++ src/album/module/view/contents/ttbcontent.cpp
@@ -30,6 +30,7 @@
 #include "controller/signalmanager.h"
 #include "imageengine/imageengineapi.h"
 #include "ac-desktop-define.h"
+#include "widgets/dialogs/imgdeletedialog.h"
 
 #include <QTimer>
 #include <QFileInfo>
@@ -1537,7 +1538,10 @@ void TTBContent::onNeedContinueRequest()
 
 void TTBContent::onTrashBtnClicked()
 {
-    emit dApp->signalM->deleteByMenu();
+    ImgDeleteDialog dialog(this, 1);
+    if (dialog.exec() == 1) {
+        emit dApp->signalM->deleteByMenu();
+    }
 }
 
 void TTBContent::onNextButton()
